# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * User registration, approval, orders, and notifications platform API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
  special: 'special',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  status: UserStatus;
  emailVerified: boolean;
  createdAt: string;
  /** @nullable */
  verificationCode?: string | null;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface VerifyEmailInput {
  code: string;
  email: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
  message?: string;
}

export type UserStatusUpdateStatus = typeof UserStatusUpdateStatus[keyof typeof UserStatusUpdateStatus];


export const UserStatusUpdateStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface UserStatusUpdate {
  status: UserStatusUpdateStatus;
  /** @nullable */
  reason?: string | null;
}

export type UserRoleUpdateRole = typeof UserRoleUpdateRole[keyof typeof UserRoleUpdateRole];


export const UserRoleUpdateRole = {
  user: 'user',
  admin: 'admin',
  special: 'special',
} as const;

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  quantity: number;
  /** @nullable */
  totalAmount?: number | null;
  status: OrderStatus;
  /** @nullable */
  shippingPartner?: string | null;
  /** @nullable */
  deliveryDate?: string | null;
  /** @nullable */
  trackingLink?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderInput {
  title: string;
  /** @nullable */
  description?: string | null;
  quantity: number;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export interface OrderShippingUpdate {
  shippingPartner: string;
  deliveryDate: string;
  trackingLink: string;
  /** @nullable */
  notes?: string | null;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  info: 'info',
  success: 'success',
  warning: 'warning',
  order_update: 'order_update',
  account_update: 'account_update',
  broadcast: 'broadcast',
} as const;

export interface Notification {
  id: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  targetRole?: string | null;
  title: string;
  message: string;
  type: NotificationType;
  isRead: boolean;
  createdAt: string;
}

export type NotificationInputType = typeof NotificationInputType[keyof typeof NotificationInputType];


export const NotificationInputType = {
  info: 'info',
  success: 'success',
  warning: 'warning',
  broadcast: 'broadcast',
} as const;

export type NotificationInputTargetType = typeof NotificationInputTargetType[keyof typeof NotificationInputTargetType];


export const NotificationInputTargetType = {
  all: 'all',
  specific: 'specific',
  special: 'special',
} as const;

export interface NotificationInput {
  title: string;
  message: string;
  type: NotificationInputType;
  targetType: NotificationInputTargetType;
  /** @nullable */
  targetUserId?: number | null;
}

export interface AdminSettings {
  acceptingOrders: boolean;
  maintenanceMode: boolean;
  updatedAt: string;
}

export interface AdminSettingsUpdate {
  acceptingOrders?: boolean;
  maintenanceMode?: boolean;
}

export interface AdminStats {
  totalUsers: number;
  pendingUsers: number;
  approvedUsers: number;
  rejectedUsers: number;
  totalOrders: number;
  pendingOrders: number;
  shippedOrders: number;
  deliveredOrders: number;
  unreadNotifications: number;
}

export type ListUsersParams = {
status?: ListUsersStatus;
role?: ListUsersRole;
};

export type ListUsersStatus = typeof ListUsersStatus[keyof typeof ListUsersStatus];


export const ListUsersStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type ListUsersRole = typeof ListUsersRole[keyof typeof ListUsersRole];


export const ListUsersRole = {
  user: 'user',
  admin: 'admin',
  special: 'special',
} as const;

export type ListUsers200 = {
  data: User[];
};

export type ListOrdersParams = {
status?: string;
userId?: number;
};

export type ListOrders200 = {
  data: Order[];
};

export type ListNotificationsParams = {
unreadOnly?: boolean;
};

export type ListNotifications200 = {
  data: Notification[];
  unreadCount: number;
};