# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: User registration, approval, orders, and notifications platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered, verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account not approved or email not verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                  unreadCount:
                    type: integer
                required:
                  - data
                  - unreadCount
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send notification (admin only) - to all, specific, or special users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get admin settings
      responses:
        "200":
          description: Admin settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update admin settings (toggle order acceptance, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin, special]
        status:
          type: string
          enum: [pending, approved, rejected]
        emailVerified:
          type: boolean
        createdAt:
          type: string
        verificationCode:
          type: ["string", "null"]
      required:
        - id
        - email
        - name
        - role
        - status
        - emailVerified
        - createdAt
    RegisterInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required:
        - email
        - password
        - name
    VerifyEmailInput:
      type: object
      properties:
        code:
          type: string
        email:
          type: string
      required:
        - code
        - email
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        message:
          type: string
      required:
        - user
        - token
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        targetRole:
          type: ["string", "null"]
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, success, warning, order_update, account_update, broadcast]
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - message
        - type
        - isRead
        - createdAt
    NotificationInput:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, success, warning, broadcast]
        targetType:
          type: string
          enum: [all, specific, special]
        targetUserId:
          type: ["integer", "null"]
      required:
        - title
        - message
        - type
        - targetType
    AdminSettings:
      type: object
      properties:
        acceptingOrders:
          type: boolean
        maintenanceMode:
          type: boolean
        updatedAt:
          type: string
      required:
        - acceptingOrders
        - maintenanceMode
        - updatedAt
    AdminSettingsUpdate:
      type: object
      properties:
        acceptingOrders:
          type: boolean
        maintenanceMode:
          type: boolean
...