# lib/api-zod/src/generated/types/adminSettingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * User registration, approval, orders, and notifications platform API
 * OpenAPI spec version: 0.2.0
 */

export interface AdminSettingsUpdate {
  acceptingOrders?: boolean;
  maintenanceMode?: boolean;
}