# lib/api-zod/src/generated/types/verifyEmailInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * User registration, approval, orders, and notifications platform API
 * OpenAPI spec version: 0.2.0
 */

export interface VerifyEmailInput {
  code: string;
  email: string;
}